Fix jar resource loading
The implementation reading the resource, as opposed to the resource as
stream, from the class loader encountered the issue that the path
derived from the resulting resource was not a valid file system path,
when executing the jar file. The path included the exclamation mark
after the jar file, pointing to the .prs file inside the root of the
jar.

Getting the resource as a stream instead returns an input stream, which
then can be read from successfully.